KAIZEN-0: Fix aktivitets visning bug
Wrappet aktiviteten inn i en route for å løse routing issuet som dukket opp
