+++ package.json
@@ -93,7 +93,7 @@
         "@craco/craco": "^5.2.3",
         "@nutgaard/use-fetch": "^2.2.1",
         "@types/classnames": "^2.2.9",
-        "@types/enzyme": "^3.10.1",
+        "@types/enzyme": "^3.10.3",
         "@types/enzyme-adapter-react-16": "^1.0.5",
         "@types/jest": "24.0.15",
         "@types/node": "^12.0.8",
@@ -104,7 +104,7 @@
         "craco-less": "^1.14.4",
         "cross-env": "^6.0.3",
         "enzyme": "^3.10.0",
-        "enzyme-adapter-react-16": "^1.14.0",
+        "enzyme-adapter-react-16": "^1.15.1",
         "enzyme-to-json": "^3.4.3",
         "husky": "^3.0.0",
         "prettier": "^1.18.2",
